// src/app/services/fetch.interface.ts
export interface Commerce {
    id:	number
    uid: string
    color: string
    department: string
    material: string
    product_name: string
    price: number
    price_string: string
    promo_code:	string
  }

export interface Computer {
    id: number
    uid: string
    platform: string
    type: string
    os: string
    stack: string
  }